# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/message-log/{uid}:
        get:
            tags:
                - MessageLog
            operationId: MessageLog_GetMessageLog
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.MessageLogItem'
    /v1/message-log/{uid}/cancel:
        put:
            tags:
                - MessageLog
            operationId: MessageLog_CancelMessage
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.CancelMessageLogRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.CancelMessageLogReply'
    /v1/message-log/{uid}/retry:
        put:
            tags:
                - MessageLog
            operationId: MessageLog_RetryMessage
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.RetryMessageLogRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.RetryMessageLogReply'
    /v1/message-logs:
        get:
            tags:
                - MessageLog
            operationId: MessageLog_ListMessageLog
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: messageType
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: startAtUnix
                  in: query
                  schema:
                    type: string
                - name: endAtUnix
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.ListMessageLogReply'
...
    /v1/sender/email/{uid}:
        post:
            tags:
                - Sender
            operationId: Sender_SendEmail
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendEmailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
    /v1/sender/email/{uid}/template:
        post:
            tags:
                - Sender
            operationId: Sender_SendEmailWithTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendEmailWithTemplateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
    /v1/sender/message:
        post:
            tags:
                - Sender
            operationId: Sender_SendMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
    /v1/sender/webhook/{uid}:
        post:
            tags:
                - Sender
            operationId: Sender_SendWebhook
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendWebhookRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
    /v1/sender/webhook/{uid}/template:
        post:
            tags:
                - Sender
            operationId: Sender_SendWebhookWithTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendWebhookWithTemplateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
...
components:
    schemas:
        rabbit.api.v1.CancelMessageLogReply:
            type: object
            properties: {}
        rabbit.api.v1.CancelMessageLogRequest:
            type: object
            properties:
                uid:
                    type: string
...
        rabbit.api.v1.ListMessageLogReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.MessageLogItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        rabbit.api.v1.MessageLogItem:
            type: object
            properties:
                uid:
                    type: string
                messageType:
                    type: integer
                    format: enum
                status:
                    type: integer
                    format: enum
                sendAt:
                    type: string
                message:
                    type: string
                config:
                    type: string
                retryTotal:
                    type: integer
                    format: int32
                lastError:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
...
        rabbit.api.v1.RetryMessageLogReply:
            type: object
            properties: {}
        rabbit.api.v1.RetryMessageLogRequest:
            type: object
            properties:
                uid:
                    type: string
...
        rabbit.api.v1.SendEmailRequest:
            type: object
            properties:
                uid:
                    type: string
                subject:
                    type: string
                body:
                    type: string
                to:
                    type: array
                    items:
                        type: string
                cc:
                    type: array
                    items:
                        type: string
                contentType:
                    type: string
                headers:
                    type: object
                    additionalProperties:
                        type: string
        rabbit.api.v1.SendEmailWithTemplateRequest:
            type: object
            properties:
                uid:
                    type: string
                templateUID:
                    type: string
                jsonData:
                    type: string
                to:
                    type: array
                    items:
                        type: string
                cc:
                    type: array
                    items:
                        type: string
        rabbit.api.v1.SendMessageRequest:
            type: object
            properties:
                uid:
                    type: string
        rabbit.api.v1.SendReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        rabbit.api.v1.SendWebhookRequest:
            type: object
            properties:
                uid:
                    type: string
                data:
                    type: string
        rabbit.api.v1.SendWebhookWithTemplateRequest:
            type: object
            properties:
                uid:
                    type: string
                templateUID:
                    type: string
                jsonData:
                    type: string
...