# pkg/api/v1/message_log_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/v1/message_log.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMessageLogCancelMessage = "/rabbit.api.v1.MessageLog/CancelMessage"
const OperationMessageLogGetMessageLog = "/rabbit.api.v1.MessageLog/GetMessageLog"
const OperationMessageLogListMessageLog = "/rabbit.api.v1.MessageLog/ListMessageLog"
const OperationMessageLogRetryMessage = "/rabbit.api.v1.MessageLog/RetryMessage"

type MessageLogHTTPServer interface {
	CancelMessage(context.Context, *CancelMessageLogRequest) (*CancelMessageLogReply, error)
	GetMessageLog(context.Context, *GetMessageLogRequest) (*MessageLogItem, error)
	ListMessageLog(context.Context, *ListMessageLogRequest) (*ListMessageLogReply, error)
	RetryMessage(context.Context, *RetryMessageLogRequest) (*RetryMessageLogReply, error)
}

func RegisterMessageLogHTTPServer(s *http.Server, srv MessageLogHTTPServer) {
	r := s.Route("/")
	r.PUT("/v1/message-log/{uid}/retry", _MessageLog_RetryMessage0_HTTP_Handler(srv))
	r.PUT("/v1/message-log/{uid}/cancel", _MessageLog_CancelMessage0_HTTP_Handler(srv))
	r.GET("/v1/message-log/{uid}", _MessageLog_GetMessageLog0_HTTP_Handler(srv))
	r.GET("/v1/message-logs", _MessageLog_ListMessageLog0_HTTP_Handler(srv))
}

func _MessageLog_RetryMessage0_HTTP_Handler(srv MessageLogHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RetryMessageLogRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageLogRetryMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RetryMessage(ctx, req.(*RetryMessageLogRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RetryMessageLogReply)
		return ctx.Result(200, reply)
	}
}

func _MessageLog_CancelMessage0_HTTP_Handler(srv MessageLogHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CancelMessageLogRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageLogCancelMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CancelMessage(ctx, req.(*CancelMessageLogRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CancelMessageLogReply)
		return ctx.Result(200, reply)
	}
}

func _MessageLog_GetMessageLog0_HTTP_Handler(srv MessageLogHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMessageLogRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageLogGetMessageLog)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMessageLog(ctx, req.(*GetMessageLogRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MessageLogItem)
		return ctx.Result(200, reply)
	}
}

func _MessageLog_ListMessageLog0_HTTP_Handler(srv MessageLogHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMessageLogRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageLogListMessageLog)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMessageLog(ctx, req.(*ListMessageLogRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMessageLogReply)
		return ctx.Result(200, reply)
	}
}

type MessageLogHTTPClient interface {
	CancelMessage(ctx context.Context, req *CancelMessageLogRequest, opts ...http.CallOption) (rsp *CancelMessageLogReply, err error)
	GetMessageLog(ctx context.Context, req *GetMessageLogRequest, opts ...http.CallOption) (rsp *MessageLogItem, err error)
	ListMessageLog(ctx context.Context, req *ListMessageLogRequest, opts ...http.CallOption) (rsp *ListMessageLogReply, err error)
	RetryMessage(ctx context.Context, req *RetryMessageLogRequest, opts ...http.CallOption) (rsp *RetryMessageLogReply, err error)
}

type MessageLogHTTPClientImpl struct {
	cc *http.Client
}

func NewMessageLogHTTPClient(client *http.Client) MessageLogHTTPClient {
	return &MessageLogHTTPClientImpl{client}
}

func (c *MessageLogHTTPClientImpl) CancelMessage(ctx context.Context, in *CancelMessageLogRequest, opts ...http.CallOption) (*CancelMessageLogReply, error) {
	var out CancelMessageLogReply
	pattern := "/v1/message-log/{uid}/cancel"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageLogCancelMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MessageLogHTTPClientImpl) GetMessageLog(ctx context.Context, in *GetMessageLogRequest, opts ...http.CallOption) (*MessageLogItem, error) {
	var out MessageLogItem
	pattern := "/v1/message-log/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageLogGetMessageLog))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MessageLogHTTPClientImpl) ListMessageLog(ctx context.Context, in *ListMessageLogRequest, opts ...http.CallOption) (*ListMessageLogReply, error) {
	var out ListMessageLogReply
	pattern := "/v1/message-logs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageLogListMessageLog))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MessageLogHTTPClientImpl) RetryMessage(ctx context.Context, in *RetryMessageLogRequest, opts ...http.CallOption) (*RetryMessageLogReply, error) {
	var out RetryMessageLogReply
	pattern := "/v1/message-log/{uid}/retry"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageLogRetryMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}