# pkg/api/v1/sender_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/v1/sender.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSenderSendEmail = "/rabbit.api.v1.Sender/SendEmail"
const OperationSenderSendEmailWithTemplate = "/rabbit.api.v1.Sender/SendEmailWithTemplate"
const OperationSenderSendMessage = "/rabbit.api.v1.Sender/SendMessage"
const OperationSenderSendWebhook = "/rabbit.api.v1.Sender/SendWebhook"
const OperationSenderSendWebhookWithTemplate = "/rabbit.api.v1.Sender/SendWebhookWithTemplate"

type SenderHTTPServer interface {
	SendEmail(context.Context, *SendEmailRequest) (*SendReply, error)
	SendEmailWithTemplate(context.Context, *SendEmailWithTemplateRequest) (*SendReply, error)
	SendMessage(context.Context, *SendMessageRequest) (*SendReply, error)
	SendWebhook(context.Context, *SendWebhookRequest) (*SendReply, error)
	SendWebhookWithTemplate(context.Context, *SendWebhookWithTemplateRequest) (*SendReply, error)
}

func RegisterSenderHTTPServer(s *http.Server, srv SenderHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/sender/message", _Sender_SendMessage1_HTTP_Handler(srv))
	r.POST("/v1/sender/email/{uid}", _Sender_SendEmail0_HTTP_Handler(srv))
	r.POST("/v1/sender/email/{uid}/template", _Sender_SendEmailWithTemplate0_HTTP_Handler(srv))
	r.POST("/v1/sender/webhook/{uid}", _Sender_SendWebhook0_HTTP_Handler(srv))
	r.POST("/v1/sender/webhook/{uid}/template", _Sender_SendWebhookWithTemplate0_HTTP_Handler(srv))
}

func _Sender_SendMessage1_HTTP_Handler(srv SenderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendMessageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSenderSendMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendMessage(ctx, req.(*SendMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendReply)
		return ctx.Result(200, reply)
	}
}

func _Sender_SendEmail0_HTTP_Handler(srv SenderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendEmailRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSenderSendEmail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendEmail(ctx, req.(*SendEmailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendReply)
		return ctx.Result(200, reply)
	}
}

func _Sender_SendEmailWithTemplate0_HTTP_Handler(srv SenderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendEmailWithTemplateRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSenderSendEmailWithTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendEmailWithTemplate(ctx, req.(*SendEmailWithTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendReply)
		return ctx.Result(200, reply)
	}
}

func _Sender_SendWebhook0_HTTP_Handler(srv SenderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendWebhookRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSenderSendWebhook)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendWebhook(ctx, req.(*SendWebhookRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendReply)
		return ctx.Result(200, reply)
	}
}

func _Sender_SendWebhookWithTemplate0_HTTP_Handler(srv SenderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendWebhookWithTemplateRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSenderSendWebhookWithTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendWebhookWithTemplate(ctx, req.(*SendWebhookWithTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendReply)
		return ctx.Result(200, reply)
	}
}

type SenderHTTPClient interface {
	SendEmail(ctx context.Context, req *SendEmailRequest, opts ...http.CallOption) (rsp *SendReply, err error)
	SendEmailWithTemplate(ctx context.Context, req *SendEmailWithTemplateRequest, opts ...http.CallOption) (rsp *SendReply, err error)
	SendMessage(ctx context.Context, req *SendMessageRequest, opts ...http.CallOption) (rsp *SendReply, err error)
	SendWebhook(ctx context.Context, req *SendWebhookRequest, opts ...http.CallOption) (rsp *SendReply, err error)
	SendWebhookWithTemplate(ctx context.Context, req *SendWebhookWithTemplateRequest, opts ...http.CallOption) (rsp *SendReply, err error)
}

type SenderHTTPClientImpl struct {
	cc *http.Client
}

func NewSenderHTTPClient(client *http.Client) SenderHTTPClient {
	return &SenderHTTPClientImpl{client}
}

func (c *SenderHTTPClientImpl) SendEmail(ctx context.Context, in *SendEmailRequest, opts ...http.CallOption) (*SendReply, error) {
	var out SendReply
	pattern := "/v1/sender/email/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSenderSendEmail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SenderHTTPClientImpl) SendEmailWithTemplate(ctx context.Context, in *SendEmailWithTemplateRequest, opts ...http.CallOption) (*SendReply, error) {
	var out SendReply
	pattern := "/v1/sender/email/{uid}/template"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSenderSendEmailWithTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SenderHTTPClientImpl) SendMessage(ctx context.Context, in *SendMessageRequest, opts ...http.CallOption) (*SendReply, error) {
	var out SendReply
	pattern := "/v1/sender/message"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSenderSendMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SenderHTTPClientImpl) SendWebhook(ctx context.Context, in *SendWebhookRequest, opts ...http.CallOption) (*SendReply, error) {
	var out SendReply
	pattern := "/v1/sender/webhook/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSenderSendWebhook))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SenderHTTPClientImpl) SendWebhookWithTemplate(ctx context.Context, in *SendWebhookWithTemplateRequest, opts ...http.CallOption) (*SendReply, error) {
	var out SendReply
	pattern := "/v1/sender/webhook/{uid}/template"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSenderSendWebhookWithTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}